Use new templating method for parsing secrets
The `secrets` specification is now processed using the new templating
method offered by Codefresh to form a single string, allowing it to be
stored in an environment variable.

Without processing the values under the `secrets`, and at the same time
passing in the value directly into the environment variable, the
resulting value of the variable will be `[object Object],[object
Object]`, given that there are two secrets specificied.

Pipeline variables are passed into the step arguments, and these
variables are referenced by the step implementation.